# apps/file-server/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VFS Link File Server API",
    "version": "1.0.0",
    "description": "HTTP API for browsing and operating on VFS Link files, uploads, shares, thumbnails, and storage-key drift. Deployments may enable HTTP Basic authentication for the entire service."
  },
  "servers": [
    {
      "url": "./",
      "description": "Current deployment and base path"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
...
  "paths": {
...
    "/api/uploads": {
      "post": {
        "tags": ["Uploads"],
        "summary": "Create an upload session",
        "description": "For local storage, uploadUrl points to this API's content endpoint. For GCS it is an opaque resumable session URI. Send each chunk with a client-generated Content-Range in addition to the fixed headers in the response.",
        "operationId": "createUpload",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUploadRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload session created, including its upload capability",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UploadResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
    "/api/uploads/{uploadId}": {
      "parameters": [
        { "$ref": "#/components/parameters/UploadId" }
      ],
      "get": {
        "tags": ["Uploads"],
        "summary": "Get an upload session",
        "description": "Returns the storage-confirmed uploadedSize and repeats the protected upload capability for active sessions so an authenticated same-origin client can resume after reload. Expired sessions have status=expired and empty capability fields.",
        "operationId": "getUpload",
        "responses": {
          "200": {
            "description": "Upload session status",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UploadResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      },
      "delete": {
        "tags": ["Uploads"],
        "summary": "Cancel an upload session",
        "operationId": "cancelUpload",
        "responses": {
          "204": { "description": "Upload session cancelled" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
    "/api/uploads/{uploadId}/content": {
      "put": {
        "tags": ["Uploads"],
        "summary": "Upload content through the file server",
        "description": "Used by local upload sessions. Send Content-Range: bytes start-end/total. start must equal uploadedSize from the latest status response. Incomplete chunks return 308 with a Range header describing committed bytes; the final chunk returns 200. GCS sessions use the same headers directly against their opaque uploadUrl.",
        "operationId": "putUploadContent",
        "parameters": [
          { "$ref": "#/components/parameters/UploadId" },
          {
            "name": "Content-Range",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "examples": ["bytes 0-8388607/16777216"] }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": { "type": "string", "format": "binary" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Final chunk accepted by the upload session",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UploadResponse" }
              }
            }
          },
          "308": {
            "description": "Chunk committed; send the next chunk after the byte identified by Range",
            "headers": {
              "Range": {
                "description": "Contiguous committed byte range, omitted when uploadedSize is zero",
                "schema": { "type": "string", "examples": ["bytes=0-8388607"] }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UploadResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "410": { "$ref": "#/components/responses/Gone" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
    "/api/uploads/{uploadId}/complete": {
      "post": {
        "tags": ["Uploads"],
        "summary": "Complete and commit an upload session",
        "description": "Call this after bytes have been uploaded either through the local content endpoint or directly to GCS.",
        "operationId": "completeUpload",
        "parameters": [
          { "$ref": "#/components/parameters/UploadId" }
        ],
        "responses": {
          "200": {
            "description": "Completed upload session",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UploadResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "410": { "$ref": "#/components/responses/Gone" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Enabled only when the deployment configures Basic authentication credentials."
      }
    },
    "parameters": {
...
      "UploadId": {
        "name": "uploadId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "minLength": 1 }
      },
...
    },
...
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Unauthorized": {
        "description": "Basic authentication is enabled and credentials are missing or invalid",
        "headers": {
          "WWW-Authenticate": { "schema": { "type": "string" } }
        },
        "content": { "text/plain": { "schema": { "type": "string", "example": "Unauthorized" } } }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "MethodNotAllowed": {
        "description": "Method not allowed",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Conflict": {
        "description": "Request conflicts with the current resource or session state",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Gone": {
        "description": "Upload session expired",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
      "ServiceUnavailable": {
        "description": "Uploads are not configured",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
    },
    "schemas": {
      "LogicalPath": {
        "type": "string",
        "description": "Slash-separated logical path relative to the root, without a leading slash.",
        "examples": ["documents/report.pdf"]
      },
      "Timestamp": {
        "type": "string",
        "format": "date-time"
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "CreateUploadRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["path", "size", "contentType", "overwrite"],
        "properties": {
          "path": { "$ref": "#/components/schemas/LogicalPath" },
          "size": { "type": "integer", "format": "int64", "minimum": 0 },
          "contentType": { "type": "string" },
          "overwrite": { "type": "boolean" }
        }
      },
      "UploadResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "logicPath", "size", "uploadedSize", "contentType", "status", "method", "uploadUrl", "headers", "completeUrl", "statusUrl", "expiresAt"],
        "properties": {
          "id": { "type": "string" },
          "logicPath": { "$ref": "#/components/schemas/LogicalPath" },
          "size": { "type": "integer", "format": "int64", "minimum": 0 },
          "uploadedSize": { "type": "integer", "format": "int64", "minimum": 0, "description": "Contiguous byte count confirmed by local staging or the GCS resumable session." },
          "contentType": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "uploading", "uploaded", "complete", "failed", "expired"] },
          "error": { "type": "string" },
          "method": { "type": "string", "const": "PUT" },
          "uploadUrl": { "type": "string", "description": "Relative local API URL or absolute opaque GCS session URI. Empty for complete or expired sessions." },
          "headers": { "type": "object", "description": "Fixed headers such as Content-Type. The client adds Content-Range for every chunk.", "additionalProperties": { "type": "string" } },
          "completeUrl": { "type": "string" },
          "statusUrl": { "type": "string" },
          "expiresAt": { "$ref": "#/components/schemas/Timestamp" }
        }
      },
...
    }
  }
}